lesson2 bugfix#2
