+++ VostrikovaE/Lesson2/Lesson2Main.java
@@ -1,0 +1,20 @@
+package VostrikovaE.Lesson2;
+
+import VostrikovaE.Lesson2.Exercise1.sumOfRandom;
+import VostrikovaE.Lesson2.Exercise2.calculateOddAndEven;
+import VostrikovaE.Lesson2.Exercise3.RandomFill;
+
+
+public class Lesson2Main {
+
+
+    public static void main(String[] args) {
+        System.out.println("exercise 1");
+        sumOfRandom.run();
+        System.out.println("exercise 2");
+        calculateOddAndEven.run();
+        System.out.println("exercise 3");
+        RandomFill.run();
+
+    }
+}
